Twitter/requirements.txt
# requirements.txt
tweepy==3.10.0
pandas==1.2.4
python-dotenv==0.17.1
PySimpleGUI==4.43.0
yagmail==0.14.256
pandasticsearch==0.5.3
elasticsearch==7.13.1
geopy==2.1.0
